+++ src/test/components/reactFCTypeEmpty/reactFCTypeEmptyResult.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function Component() {
+  return <Extracted />;
+}
+
+const Extracted: React.FC = () => (
+  <div className='my-class-2'>
+    <div className='my-class'>Test</div>
+    <span className={'string-literal'}></span>
+  </div>
+);
+